Upgrade to 1.1.7 in order to fix issues with the latest version of LiblineaR
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/kebabs@99221 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/MotifC.cpp
@@ -75,7 +75,7 @@ struct intfFindMotifs {
     int offset;
     int rowIndex;
     int numUsedMotifs;
-    int numNonzeroFeatures;
+    uint64_t numNonzeroFeatures;
     int featuresPerSample;
     int svmIndex;
     char *keyPool;
